Update deleteAllVnets script to take into account the vnet age
