+++ package.json
@@ -33,8 +33,8 @@
     "vnet:fork": "node --env-file=.env ./scripts/forkVnet.ts",
     "vnet:fork:ci": "node ./scripts/forkVnet.ts",
     "vnet:delete": "node --env-file=.env ./scripts/deleteVnet.ts",
-    "vnet:delete:all": "node --env-file=.env ./scripts/deleteAllVnetsNewEndpoint.js 'local-tests-testnet'",
-    "vnet:delete:all:ci": "node --env-file=.env ./scripts/deleteAllVnetsNewEndpoint.js 'ci-tests-testnet'",
+    "vnet:delete:all": "node --env-file=.env ./scripts/deleteAllVnets.js 'local-tests-testnet'",
+    "vnet:delete:all:ci": "node --env-file=.env ./scripts/deleteAllVnets.js 'ci-tests-testnet'",
     "vnet:delete:ci": "node ./scripts/deleteVnet.ts",
     "prepare": "husky",
     "typecheck": "pnpm --filter './packages/**' --filter './apps/**' typecheck",
